Move the ereader-metadata read into library's queries package
The last read left in `library/use_cases/` is a single
`find_by_client_book_id` plus a projection onto a small DTO. It takes the
ADR's halfway option at its most minimal: the use case relocates as-is,
keeps delegating to `BookRepository` and keeps returning
`EreaderMetadata`. A full port would remove nothing, since the finder is
shared with the create-book and ebook-upload commands.

The point is the package placement — `use_cases/` is now commands-only —
not ceremony for a lightweight read.
