--- backend/src/application/library/use_cases/book_queries/__init__.py
@@ -1,1 +1,0 @@
-"""Book query use cases (read-only listing and metadata queries)."""
